+++ packages/svelte/src/lib/components/Authenticator.svelte
@@ -1,52 +1,52 @@
 <script lang="ts">
-	import { onDestroy } from 'svelte';
-	import { setupMachine, route, toSignIn } from '$lib/components/store';
-	import SignIn from './SignIn.svelte';
-	import AmplifyTabs from './primitives/AmplifyTabs.svelte';
+  import { onDestroy } from 'svelte';
+  import { setupMachine, route, toSignIn } from '$lib/components/authStore';
+  import SignIn from './SignIn.svelte';
+  import AmplifyTabs from './primitives/AmplifyTabs.svelte';
 
-	export let initialState = undefined;
-	export let loginMechanisms = undefined;
-	export let services = undefined;
-	export let signUpAttributes = undefined;
-	export let socialProviders = undefined;
+  export let initialState = undefined;
+  export let loginMechanisms = undefined;
+  export let services = undefined;
+  export let signUpAttributes = undefined;
+  export let socialProviders = undefined;
 
-	const _subscription = setupMachine(
-		initialState,
-		loginMechanisms,
-		services,
-		signUpAttributes,
-		socialProviders
-	);
+  const _subscription = setupMachine(
+    initialState,
+    loginMechanisms,
+    services,
+    signUpAttributes,
+    socialProviders
+  );
 
-	onDestroy(() => {
-		if (_subscription) _subscription.unsubscribe();
-	});
+  onDestroy(() => {
+    if (_subscription) _subscription.unsubscribe();
+  });
 </script>
 
 {#if $route !== 'authenticated'}
-	<div data-amplify-authenticator>
-		<div data-amplify-container>
-			<!--Slot header goes here-->
-			<div data-amplify-router>
-				{#if $route === 'signIn' || $route === 'signUp'}
-					<AmplifyTabs />
-				{/if}
+  <div data-amplify-authenticator>
+    <div data-amplify-container>
+      <!--Slot header goes here-->
+      <div data-amplify-router>
+        {#if $route === 'signIn' || $route === 'signUp'}
+          <AmplifyTabs />
+        {/if}
 
-				{#if $route === 'signIn'}
-					<SignIn />
-				{/if}
-				{#if $route === 'signUp'}
-					<p>Sign Up</p>
-					<button on:click={toSignIn}>To Sign In</button>
-				{/if}
-				{#if $route === 'resetPassword'}
-					<p>reset password</p>
-					<button on:click={toSignIn}>To Sign In</button>
-				{/if}
-			</div>
-		</div>
-	</div>
+        {#if $route === 'signIn'}
+          <SignIn />
+        {/if}
+        {#if $route === 'signUp'}
+          <p>Sign Up</p>
+          <button on:click={toSignIn}>To Sign In</button>
+        {/if}
+        {#if $route === 'resetPassword'}
+          <p>reset password</p>
+          <button on:click={toSignIn}>To Sign In</button>
+        {/if}
+      </div>
+    </div>
+  </div>
 {/if}
 {#if $route === 'authenticated'}
-	<slot />
+  <slot />
 {/if}
